Allow pulling images from private Docker registry
As per the documentation for Fargate using private docker registries[1]
this commit introduces a new variable
`repository_credentials_secret_arn` used to pass in the AWS Secrets
Manager path to your credentials secret. Note that this also requires
the Lambda function execution task role to have access to this secret.

[1] https://aws.amazon.com/blogs/compute/introducing-private-registry-authentication-support-for-aws-fargate/